Add shift transaction entity within graphql server into individual files
