+++ packages/mock-server/src/schema/Transaction/index.js
@@ -1,0 +1,16 @@
+import {
+  TransactionInput,
+  TransactionMutationResolvers,
+  TransactionMutations,
+} from './mutations';
+import { TransactionTypes } from './types';
+import { TransactionQueries, TransactionQueryResolvers } from './queries';
+
+export {
+  TransactionMutations,
+  TransactionTypes,
+  TransactionQueryResolvers,
+  TransactionQueries,
+  TransactionMutationResolvers,
+  TransactionInput,
+};
