+++ packages/mock-server/src/schema/Transaction/types.js
@@ -1,0 +1,13 @@
+export const TransactionTypes = `
+    type Transaction {
+        id: String
+        date: String
+        customer: String
+        supplier: String
+        total: String
+    }
+    type TransactionResponse { 
+      data: [Transaction],
+      totalLength: Int
+    }
+  `;
